refactor(app): delete legacy chart-registry, migrate to chart-helpers (Phase 7)
Replace all imports from @/lib/chart-registry with @/lib/chart-helpers.
The new module delegates to pluginRegistry and includes lightweight
plugin registration for test environments.

- Create app/src/lib/chart-helpers.ts with helper functions
- Create app/src/lib/__tests__/chart-helpers.test.ts with 21 tests
- Migrate 15 consumer files from chart-registry to chart-helpers
- Delete chart-registry.ts and its 3 test files
- Update test mocks to include getChartOptions for plugin imports
