+++ app/src/lib/__tests__/chart-helpers.test.ts
@@ -1,0 +1,195 @@
+import { describe, it, expect, vi } from "vitest";
+
+const { Stub } = vi.hoisted(() => ({ Stub: () => null }));
+
+vi.mock("@neoboard/components", () => ({
+  getChartOptions: () => [],
+  BarChart: Stub,
+  LineChart: Stub,
+  PieChart: Stub,
+  SingleValueChart: Stub,
+  GraphChart: Stub,
+  MapChart: Stub,
+  JsonViewer: Stub,
+  MarkdownWidget: Stub,
+  IframeWidget: Stub,
+  GaugeChart: Stub,
+  SankeyChart: Stub,
+  SunburstChart: Stub,
+  RadarChart: Stub,
+  TreemapChart: Stub,
+  EmptyState: Stub,
+  Skeleton: Stub,
+}));
+
+vi.mock("@/components/table-renderer", () => ({ TableRenderer: Stub }));
+vi.mock("@/components/parameter-widget-renderer", () => ({
+  ParameterWidgetRenderer: Stub,
+}));
+vi.mock("@/components/form-widget-renderer", () => ({
+  FormWidgetRenderer: Stub,
+}));
+vi.mock("@/components/graph-exploration-wrapper", () => ({
+  GraphExplorationWrapper: Stub,
+}));
+
+// Ensure plugins are registered before importing helpers.
+import "@/plugins/index";
+import {
+  getChartConfig,
+  chartSupportsClickAction,
+  chartSupportsStyling,
+  getStylingTargets,
+  getCompatibleChartTypes,
+  chartRequiresQuery,
+  getChartDefaults,
+  supportsColumnMapping,
+  getAllChartTypes,
+  CHART_TYPES,
+} from "../chart-helpers";
+
+// ---------------------------------------------------------------------------
+// getChartConfig
+// ---------------------------------------------------------------------------
+describe("getChartConfig", () => {
+  it("returns a plugin for a registered type", () => {
+    const config = getChartConfig("bar");
+    expect(config).toBeDefined();
+    expect(config!.type).toBe("bar");
+    expect(config!.label).toBe("Bar Chart");
+  });
+
+  it("returns undefined for an unknown type", () => {
+    expect(getChartConfig("nonexistent")).toBeUndefined();
+  });
+});
+
+// ---------------------------------------------------------------------------
+// chartSupportsClickAction
+// ---------------------------------------------------------------------------
+describe("chartSupportsClickAction", () => {
+  it("returns true for bar (supports click)", () => {
+    expect(chartSupportsClickAction("bar")).toBe(true);
+  });
+
+  it("returns false for single-value (no click)", () => {
+    expect(chartSupportsClickAction("single-value")).toBe(false);
+  });
+
+  it("returns false for unknown types", () => {
+    expect(chartSupportsClickAction("unknown")).toBe(false);
+  });
+});
+
+// ---------------------------------------------------------------------------
+// chartSupportsStyling
+// ---------------------------------------------------------------------------
+describe("chartSupportsStyling", () => {
+  it("returns true for bar", () => {
+    expect(chartSupportsStyling("bar")).toBe(true);
+  });
+
+  it("returns false for json", () => {
+    expect(chartSupportsStyling("json")).toBe(false);
+  });
+
+  it("returns false for unknown types", () => {
+    expect(chartSupportsStyling("unknown")).toBe(false);
+  });
+});
+
+// ---------------------------------------------------------------------------
+// getStylingTargets
+// ---------------------------------------------------------------------------
+describe("getStylingTargets", () => {
+  it("returns targets for bar chart", () => {
+    const targets = getStylingTargets("bar");
+    expect(targets.length).toBeGreaterThan(0);
+    expect(targets[0]).toHaveProperty("value");
+    expect(targets[0]).toHaveProperty("label");
+  });
+
+  it("returns empty array for json (no styling)", () => {
+    expect(getStylingTargets("json")).toEqual([]);
+  });
+
+  it("returns empty array for unknown types", () => {
+    expect(getStylingTargets("unknown")).toEqual([]);
+  });
+});
+
+// ---------------------------------------------------------------------------
+// getCompatibleChartTypes
+// ---------------------------------------------------------------------------
+describe("getCompatibleChartTypes", () => {
+  it("returns all types for neo4j", () => {
+    const types = getCompatibleChartTypes("neo4j");
+    expect(types).toContain("bar");
+    expect(types).toContain("graph");
+  });
+
+  it("returns types excluding graph for postgresql", () => {
+    const types = getCompatibleChartTypes("postgresql");
+    expect(types).toContain("bar");
+    expect(types).not.toContain("graph");
+  });
+
+  it("returns empty array for invalid connector type", () => {
+    expect(getCompatibleChartTypes("invalid")).toEqual([]);
+  });
+});
+
+// ---------------------------------------------------------------------------
+// chartRequiresQuery
+// ---------------------------------------------------------------------------
+describe("chartRequiresQuery", () => {
+  it("returns true for bar", () => {
+    expect(chartRequiresQuery("bar")).toBe(true);
+  });
+
+  it("returns false for markdown", () => {
+    expect(chartRequiresQuery("markdown")).toBe(false);
+  });
+
+  it("defaults to true for unknown types", () => {
+    expect(chartRequiresQuery("unknown")).toBe(true);
+  });
+});
+
+// ---------------------------------------------------------------------------
+// getChartDefaults
+// ---------------------------------------------------------------------------
+describe("getChartDefaults", () => {
+  it("returns empty object", () => {
+    expect(getChartDefaults("bar")).toEqual({});
+  });
+});
+
+// ---------------------------------------------------------------------------
+// supportsColumnMapping
+// ---------------------------------------------------------------------------
+describe("supportsColumnMapping", () => {
+  it("returns true for bar, line, pie", () => {
+    expect(supportsColumnMapping("bar")).toBe(true);
+    expect(supportsColumnMapping("line")).toBe(true);
+    expect(supportsColumnMapping("pie")).toBe(true);
+  });
+
+  it("returns false for table, json", () => {
+    expect(supportsColumnMapping("table")).toBe(false);
+    expect(supportsColumnMapping("json")).toBe(false);
+  });
+});
+
+// ---------------------------------------------------------------------------
+// getAllChartTypes
+// ---------------------------------------------------------------------------
+describe("getAllChartTypes", () => {
+  it("returns all 17 registered types", () => {
+    const types = getAllChartTypes();
+    expect(types.length).toBe(17);
+    for (const t of CHART_TYPES) {
+      expect(types).toContain(t);
+    }
+  });
+});
