+++ app/src/lib/chart-helpers.ts
@@ -1,0 +1,300 @@
+/**
+ * Chart helper functions — thin wrappers around the plugin registry.
+ *
+ * These replace the legacy chart-registry.ts helpers with direct
+ * delegation to `pluginRegistry`. Consumers that previously imported
+ * from `@/lib/chart-registry` should import from here instead.
+ *
+ * Importing this module has the side effect of registering lightweight
+ * (stub) plugins when the full plugin modules haven't loaded yet. At
+ * runtime, `plugins/index.ts` replaces these stubs with real
+ * components; but for tests that never import full plugins, the stubs
+ * provide enough metadata for helper functions to work.
+ */
+
+import { pluginRegistry } from "@/plugins/registry";
+import { defineChartPlugin } from "@/lib/chart-plugin-registry";
+import type { ChartPlugin } from "@/lib/chart-plugin-registry";
+import { CONNECTOR_TYPES, type ConnectorType } from "@/lib/connector-types";
+
+// Re-export types for backward compatibility
+export type { ChartType } from "@/plugins/chart-types";
+export { CHART_TYPES } from "@/plugins/chart-types";
+export type { ConnectorType } from "@/lib/connector-types";
+export type { ColumnMapping } from "@neoboard/components";
+export type { ChartPlugin };
+
+// ---------------------------------------------------------------------------
+// Column mapping support — hardcoded set of types that support it.
+// ---------------------------------------------------------------------------
+
+const COLUMN_MAPPING_TYPES = new Set<string>(["bar", "line", "pie"]);
+
+// ---------------------------------------------------------------------------
+// Lightweight plugin registration (no React component imports)
+//
+// Ensures that helper functions return correct capability information
+// even before the full plugin modules (with React components) load.
+// At app startup, `plugins/index.ts` replaces stubs with real plugins.
+// ---------------------------------------------------------------------------
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any -- stub placeholder
+const StubComponent: React.ComponentType<any> = (() => null) as any;
+const identity = (d: unknown) => d;
+const nullTransform = () => null;
+
+interface LightDef {
+  type: string;
+  label: string;
+  compatibleWith?: ConnectorType[];
+  stylingTargets?: { value: string; label: string }[];
+  capabilities?: {
+    supportsClickAction?: boolean;
+    supportsStyling?: boolean;
+    isECharts?: boolean;
+    requiresQuery?: boolean;
+  };
+}
+
+const LIGHTWEIGHT_DEFS: LightDef[] = [
+  {
+    type: "bar",
+    label: "Bar Chart",
+    compatibleWith: ["neo4j", "postgresql"],
+    stylingTargets: [{ value: "color", label: "Bar Color" }],
+    capabilities: { isECharts: true, supportsStyling: true },
+  },
+  {
+    type: "line",
+    label: "Line Chart",
+    compatibleWith: ["neo4j", "postgresql"],
+    stylingTargets: [{ value: "color", label: "Line Color" }],
+    capabilities: { isECharts: true, supportsStyling: true },
+  },
+  {
+    type: "pie",
+    label: "Pie / Doughnut",
+    compatibleWith: ["neo4j", "postgresql"],
+    stylingTargets: [{ value: "color", label: "Slice Color" }],
+    capabilities: { isECharts: true, supportsStyling: true },
+  },
+  {
+    type: "table",
+    label: "Data Table",
+    compatibleWith: ["neo4j", "postgresql"],
+    stylingTargets: [
+      { value: "backgroundColor", label: "Background Color" },
+      { value: "textColor", label: "Text Color" },
+    ],
+    capabilities: { supportsStyling: true },
+  },
+  {
+    type: "single-value",
+    label: "Single Value",
+    compatibleWith: ["neo4j", "postgresql"],
+    stylingTargets: [
+      { value: "color", label: "Text Color" },
+      { value: "backgroundColor", label: "Background Color" },
+    ],
+    capabilities: {
+      supportsClickAction: false,
+      isECharts: true,
+      supportsStyling: true,
+    },
+  },
+  {
+    type: "graph",
+    label: "Graph",
+    compatibleWith: ["neo4j"],
+    stylingTargets: [{ value: "color", label: "Node Color" }],
+    capabilities: { supportsStyling: true },
+  },
+  {
+    type: "map",
+    label: "Map",
+    compatibleWith: ["neo4j", "postgresql"],
+    stylingTargets: [{ value: "color", label: "Marker Color" }],
+    capabilities: { supportsStyling: true },
+  },
+  {
+    type: "json",
+    label: "JSON Viewer",
+    compatibleWith: ["neo4j", "postgresql"],
+    capabilities: { supportsClickAction: false, supportsStyling: false },
+  },
+  {
+    type: "parameter-select",
+    label: "Parameter Selector",
+    compatibleWith: ["neo4j", "postgresql"],
+    capabilities: {
+      supportsClickAction: false,
+      supportsStyling: false,
+      requiresQuery: false,
+    },
+  },
+  {
+    type: "form",
+    label: "Form",
+    compatibleWith: ["neo4j", "postgresql"],
+    capabilities: { supportsStyling: false, requiresQuery: false },
+  },
+  {
+    type: "markdown",
+    label: "Markdown",
+    compatibleWith: ["neo4j", "postgresql"],
+    capabilities: {
+      supportsClickAction: false,
+      supportsStyling: false,
+      requiresQuery: false,
+    },
+  },
+  {
+    type: "iframe",
+    label: "iFrame",
+    compatibleWith: ["neo4j", "postgresql"],
+    capabilities: {
+      supportsClickAction: false,
+      supportsStyling: false,
+      requiresQuery: false,
+    },
+  },
+  {
+    type: "gauge",
+    label: "Gauge",
+    compatibleWith: ["neo4j", "postgresql"],
+    stylingTargets: [{ value: "color", label: "Gauge Color" }],
+    capabilities: { isECharts: true, supportsStyling: true },
+  },
+  {
+    type: "sankey",
+    label: "Sankey",
+    compatibleWith: ["neo4j", "postgresql"],
+    stylingTargets: [{ value: "color", label: "Link Color" }],
+    capabilities: { isECharts: true, supportsStyling: true },
+  },
+  {
+    type: "sunburst",
+    label: "Sunburst",
+    compatibleWith: ["neo4j", "postgresql"],
+    stylingTargets: [{ value: "color", label: "Segment Color" }],
+    capabilities: { isECharts: true, supportsStyling: true },
+  },
+  {
+    type: "radar",
+    label: "Radar",
+    compatibleWith: ["neo4j", "postgresql"],
+    stylingTargets: [{ value: "color", label: "Area Color" }],
+    capabilities: {
+      supportsClickAction: false,
+      isECharts: true,
+      supportsStyling: true,
+    },
+  },
+  {
+    type: "treemap",
+    label: "Treemap",
+    compatibleWith: ["neo4j", "postgresql"],
+    stylingTargets: [{ value: "color", label: "Block Color" }],
+    capabilities: { isECharts: true, supportsStyling: true },
+  },
+];
+
+for (const def of LIGHTWEIGHT_DEFS) {
+  if (!pluginRegistry.has(def.type)) {
+    pluginRegistry.register(
+      defineChartPlugin({
+        type: def.type,
+        label: def.label,
+        component: StubComponent,
+        transform:
+          def.type === "markdown" || def.type === "iframe"
+            ? nullTransform
+            : identity,
+        transformWithMapping:
+          def.type === "markdown" || def.type === "iframe"
+            ? nullTransform
+            : identity,
+        compatibleWith: def.compatibleWith,
+        stylingTargets: def.stylingTargets,
+        capabilities: def.capabilities,
+      }),
+    );
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Helper functions
+// ---------------------------------------------------------------------------
+
+/**
+ * Get a chart plugin by type. Returns undefined if not registered.
+ */
+export function getChartConfig(type: string): ChartPlugin | undefined {
+  return pluginRegistry.get(type);
+}
+
+/**
+ * Check whether a chart type supports click actions.
+ */
+export function chartSupportsClickAction(type: string): boolean {
+  return pluginRegistry.get(type)?.capabilities.supportsClickAction ?? false;
+}
+
+/**
+ * Check whether a chart type supports rule-based styling.
+ */
+export function chartSupportsStyling(type: string): boolean {
+  return pluginRegistry.get(type)?.capabilities.supportsStyling ?? false;
+}
+
+/**
+ * Get the styling targets for a chart type.
+ * Returns an empty array if the type doesn't support styling.
+ */
+export function getStylingTargets(
+  type: string,
+): { value: string; label: string }[] {
+  const plugin = pluginRegistry.get(type);
+  if (!plugin?.capabilities.supportsStyling) return [];
+  return plugin.stylingTargets ?? [{ value: "color", label: "Color" }];
+}
+
+/**
+ * Get all chart types compatible with a given connector type.
+ */
+export function getCompatibleChartTypes(connectorType: string): string[] {
+  if (!CONNECTOR_TYPES.includes(connectorType as ConnectorType)) return [];
+  return pluginRegistry
+    .getCompatibleWith(connectorType as ConnectorType)
+    .map((p) => p.type);
+}
+
+/**
+ * Check whether a chart type requires a query to render.
+ */
+export function chartRequiresQuery(type: string): boolean {
+  return pluginRegistry.get(type)?.capabilities.requiresQuery ?? true;
+}
+
+/**
+ * Get default chart settings for a chart type.
+ * Returns an empty object — defaults are managed by Zod schemas in plugins.
+ */
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export function getChartDefaults(_type: string): Record<string, unknown> {
+  return {};
+}
+
+/**
+ * Check whether a chart type supports column mapping.
+ */
+export function supportsColumnMapping(type: string): boolean {
+  return COLUMN_MAPPING_TYPES.has(type);
+}
+
+/**
+ * Get all registered chart types as strings.
+ */
+export function getAllChartTypes(): string[] {
+  return pluginRegistry.getTypes();
+}
