+++ app/src/plugins/__tests__/bar.test.tsx
@@ -18,6 +18,7 @@ vi.mock("next/dynamic", () => ({
 
 vi.mock("@neoboard/components", () => ({
   Skeleton: () => null,
+  getChartOptions: () => [],
 }));
 
 describe("barPlugin", () => {
